pyproject.toml
[project]
name = "radon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pynacl>=1.5.0",
    "websockets>=15.0.1",
]
